requirements.txt
faker==19.13.0
psycopg2-binary==2.9.9
kafka-python==2.0.2
pandas==1.5.3
numpy==1.24.4
requests==2.31.0 
google-cloud-bigquery==3.11.4
dbt-bigquery==1.7.0